requirements.txt
streamlit>=1.35.0
pyyaml>=6.0
aiohttp>=3.9.0
